text2img/requirements.txt
stable-diffusion-cpp-python